Cleaned up.
Stopping the ball is better. We now look for linear momentum to approach
zero as well as linear velocity, then we stop and hold it.

Now we store strokes per level and then display them on the form too.
They store!!!!

I want elastic collisions. So bad, that it hurts.